requirements.txt
blurhash==1.1.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Mastodon.py==1.6.3
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
uWSGI==2.0.21
Werkzeug==2.2.2
zipp==3.10.0
